SA: store suffix/LCP arrays as u32, halving the per-fragment floor
The SA values come from 32-bit libsais (i32, text < 2^31) but were widened to
u64 for suftab/lcptab — pure waste. Narrow them to u32: the values provably fit,
so it's a numerical no-op, and it halves the largest fixed memory term (~160 ->
~80 MB per 5 Mbp fragment). enumerate_maxpairs now takes &[u32] and widens to u64
at its two read sites; all internal arithmetic stays u64.

tirvish_rs targets pre-fragmented input, so the mirrored text is always far under
2^31. Added an explicit guard (panics with a helpful message if a single input
file exceeds the 32-bit limit, rather than corrupting) and a README "Input size"
caveat documenting the pre-fragmentation expectation.

Bit-exact on the testdata oracle: 270/267/295/321, all 14 columns.

+++ src/maxpairs.rs
@@ -273,8 +273,8 @@ struct ItvInfo {
 /// (`lcptab[0] == 0`); `n = suftab.len()` = number of non-special suffixes.
 /// `enc` = per-position left-char codes (specials == `alphabetsize`).
 pub fn enumerate_maxpairs(
-    suftab: &[u64],
-    lcptab: &[u64],
+    suftab: &[u32],
+    lcptab: &[u32],
     searchlength: u64,
     alphabetsize: u32,
     enc: &[u32],
@@ -294,12 +294,12 @@ pub fn enumerate_maxpairs(
     let mut lastinterval: Option<ItvInfo> = None;
 
     for idx in 0..n {
-        let previoussuffix = suftab[idx];
+        let previoussuffix = suftab[idx] as u64;
         // LCP is LOOK-AHEAD: lcp(suftab[idx], suftab[idx+1]) == standard
         // look-behind lcptab[idx+1] (0 past the end). This is gt's streaming
         // convention; using look-behind lcptab[idx] drops the first leaf of
         // each interval. Verified against a brute-force maximal-pairs oracle.
-        let lcpvalue = if idx + 1 < n { lcptab[idx + 1] } else { 0 };
+        let lcpvalue = if idx + 1 < n { lcptab[idx + 1] as u64 } else { 0 };
 
         if lcpvalue <= stack.last().unwrap().lcp {
             let firstedge;
